refactor DM3 and 4 code to use the shared DM base class and moved it into the DM_IO namespace
